pyproject.toml
[project]
name = "autoriff"
version = "0.1.0"
description = "Temporal AI Agent for Iterative Idea Development"
requires-python = ">=3.11"
dependencies = [
    "temporalio>=1.8.0",
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
openai = [
    "openai>=1.0.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["models", "prompts", "activities", "workflows", "cli", "scripts"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
